+++ node-graph/interpreted-executor/Cargo.toml
@@ -27,3 +27,4 @@ serde = { version = "1", features = ["derive"], optional = true }
 glam = { version = "0.22" }
 once_cell = "1.17.0"
 futures = "0.3.28"
+typed-arena = "2.0.2"
