+++ node-graph/interpreted-executor/src/lib.rs
@@ -1,7 +1,7 @@
 #[macro_use]
 extern crate log;
 
-pub mod executor;
+pub mod dynamic_executor;
 pub mod node_registry;
 
 #[cfg(test)]
@@ -72,8 +72,8 @@ mod tests {
 			..Default::default()
 		};
 
-		use crate::executor::DynamicExecutor;
-		use graph_craft::executor::{Compiler, Executor};
+		use crate::dynamic_executor::DynamicExecutor;
+		use graph_craft::graphene_compiler::{Compiler, Executor};
 
 		let compiler = Compiler {};
 		let protograph = compiler.compile_single(network, true).expect("Graph should be generated");
@@ -120,8 +120,8 @@ mod tests {
 			..Default::default()
 		};
 
-		use crate::executor::DynamicExecutor;
-		use graph_craft::executor::Compiler;
+		use crate::dynamic_executor::DynamicExecutor;
+		use graph_craft::graphene_compiler::Compiler;
 
 		let compiler = Compiler {};
 		let protograph = compiler.compile_single(network, true).expect("Graph should be generated");
